location update

+++ app/src/main/java/com/example/weather/dataclass/data/currentweather/Coord.kt
@@ -1,6 +1,6 @@
 package com.example.weather.dataclass.data.currentweather
 
 data class Coord(
-    val lat: Int?,
-    val lon: Int?
+    val lat: Float?,
+    val lon: Float?
 )
